edit delete blog

+++ controllers/blogs.js
@@ -57,7 +57,23 @@ blogRouter.get('/:id', async(req, res) => {
 })
 
 blogRouter.delete('/:id', async(req, res) =>{
-    await Blog.findByIdAndDelete(req.params.id)
+    const token = req.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!token || !decodedToken.id){
+        return res.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(req.params.id)
+
+    if (blog.user.toString() !== user.id.toString())
+        return res.status(401).json({ error: 'unauthorized user' })
+
+    Blog.findByIdAndDelete(req.params.id)
+    user.blogs = user.blogs.filter(blog => blog.id.toString() !== req.params.id.toString())
+    await user.save()
+
     res.status(204).end()
 })
